+++ src/system.cpp
@@ -32,8 +32,8 @@ void System::initWaterGrid() {
             for (int k = 0; k < WATERGRID_Z; k++) {
                 /// Create the cell
                 Cell cell {
-                    .oldVelocity  = Vector3f(0, 0, 0), // CUSTOMIZABLE
-                    .currVelocity = Vector3f(0, 0, 0), // CUSTOMIZABLE
+                    .oldVelocity  = Vector3d(0, 0, 0), // CUSTOMIZABLE
+                    .currVelocity = Vector3d(0, 0, 0), // CUSTOMIZABLE
                     .forceApplied = false,
                     .neighbors = getGridNeighbors(i, j, k)
                 };
@@ -43,8 +43,8 @@ void System::initWaterGrid() {
             }
         }
     }
-    float product = WATERGRID_X*WATERGRID_Z*WATERGRID_Y;
-    float size = m_waterGrid.size();
+    double product = WATERGRID_X*WATERGRID_Z*WATERGRID_Y;
+    double size = m_waterGrid.size();
     assert(size == WATERGRID_X*WATERGRID_Z*WATERGRID_Y);
     initPressureA();
 }
@@ -56,17 +56,17 @@ int nonZeroRand() {
 }
 
 /// Returns a random position within the specified ranges
-Vector3f getRandPosWithinRange(float minX, float maxX,
-                               float minY, float maxY,
-                               float minZ, float maxZ) {
-    float x = minX + ((maxX - minX) / (nonZeroRand() % 100 + 1));
+Vector3d getRandPosWithinRange(double minX, double maxX,
+                               double minY, double maxY,
+                               double minZ, double maxZ) {
+    double x = minX + ((maxX - minX) / (nonZeroRand() % 100 + 1));
     assert(!isinf(x) || !isnan(x));
-    float y = minY + ((maxY - minY) / (nonZeroRand() % 100 + 1));
+    double y = minY + ((maxY - minY) / (nonZeroRand() % 100 + 1));
     assert(!isinf(y) || !isnan(y));
-    float z = minZ + ((maxZ - minZ) / (nonZeroRand() % 100 + 1));
+    double z = minZ + ((maxZ - minZ) / (nonZeroRand() % 100 + 1));
     assert(!isinf(z) || !isnan(z));
 
-    return Vector3f(x, y, z);
+    return Vector3d(x, y, z);
 }
 
 /// Initializes INIT_NUM_PARTICLES Particle structs
@@ -78,7 +78,7 @@ void System::initParticles() {
             .position = getRandPosWithinRange(WATERGRID_X/2.f-0.2, WATERGRID_X/2.f+0.2,
                                               WATERGRID_Y - 0.1, WATERGRID_Y - 0.1,
                                               WATERGRID_Z/2.f-0.2, WATERGRID_Z/2.f+0.2), // CUSTOMIZABLE
-            .velocity = Vector3f{0, 0, 0}, // CUSTOMIZABLE
+            .velocity = Vector3d{0, 0, 0}, // CUSTOMIZABLE
             .opacity  = 1.f,
             .lifeTime = 5.f // CUSTOMIZABLE
         };
